% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{AGENCY}
\alias{AGENCY}
\title{a list of OD data agency}
\format{
An object of class \code{character} of length 17.
}
\usage{
AGENCY
}
\description{
a list of OD data agency
}
\keyword{datasets}
