% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-MapSf.R
\docType{data}
\name{COLUMBUS_STATIC_ROADMAP}
\alias{COLUMBUS_STATIC_ROADMAP}
\title{A static Columbus roadmap for opioid hot spot detection}
\format{
An object of class \code{ggmap} (inherits from \code{raster}) with 1280 rows and 1280 columns.
}
\usage{
COLUMBUS_STATIC_ROADMAP
}
\description{
A static Columbus roadmap for opioid hot spot detection
}
\keyword{datasets}
