% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-drugCrime.R
\name{data_drug_crime}
\alias{data_drug_crime}
\title{Loading function for drug crime data}
\usage{
data_drug_crime(parquet = FALSE)
}
\arguments{
\item{parquet}{parquet format}
}
\value{
drug crime data
}
\description{
Loading function for drug crime data
}
