% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-overdoseData.R
\name{data_opioid_overdose}
\alias{data_opioid_overdose}
\title{Load opioid overdose primary table}
\usage{
data_opioid_overdose()
}
\value{
opioid overdose data in primary table format
}
\description{
Load opioid overdose primary table
}
