% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_crime_data_loading_functions.R
\name{drug_crime_data}
\alias{drug_crime_data}
\title{Loading function for drug crime data}
\usage{
drug_crime_data(
  dir = opioidDashboard::DRUG_CRIME_DATA_DIRECTORY,
  geocode = TRUE
)
}
\arguments{
\item{dir}{directory on OSC that contains drug crime data}

\item{geocode}{TRUE will return data with lat/lng geocoding information}
}
\value{
drug crime data
}
\description{
Loading function for drug crime data
}
\examples{
drug_crime_data()
}
