% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-MapSf.R
\docType{data}
\name{ELIGIBLE_BUSINESS_TYPE}
\alias{ELIGIBLE_BUSINESS_TYPE}
\title{Eligible business type for hot spot detection}
\format{
An object of class \code{character} of length 11.
}
\usage{
ELIGIBLE_BUSINESS_TYPE
}
\description{
Eligible business type for hot spot detection
}
\keyword{datasets}
