% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-MapSf.R
\docType{data}
\name{FCPH_LOCATION_OF_INTEREST}
\alias{FCPH_LOCATION_OF_INTEREST}
\title{FCPH_LOCATION_OF_INTEREST}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 10 columns.
}
\usage{
FCPH_LOCATION_OF_INTEREST
}
\description{
FCPH_LOCATION_OF_INTEREST
}
\keyword{datasets}
