% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter_selection.R
\docType{data}
\name{filter_selection_destination}
\alias{filter_selection_destination}
\title{Filter selection for destination}
\format{
An object of class \code{character} of length 25.
}
\usage{
filter_selection_destination
}
\description{
Filter selection for destination
}
\keyword{datasets}
