% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overdose_filter_selection.R
\docType{data}
\name{filter_selection_location_type}
\alias{filter_selection_location_type}
\title{Filter selection for location type}
\format{
An object of class \code{character} of length 13.
}
\usage{
filter_selection_location_type
}
\description{
Filter selection for location type
}
\keyword{datasets}
