% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataFilterSelection.R
\docType{data}
\name{filter_selection_treatment_providers_service}
\alias{filter_selection_treatment_providers_service}
\title{Filter selection for treatment providers service type}
\format{
An object of class \code{character} of length 79.
}
\usage{
filter_selection_treatment_providers_service
}
\description{
Filter selection for treatment providers service type
}
\keyword{datasets}
