% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter_selection.R
\docType{data}
\name{filter_selection_treatment_providers_spec}
\alias{filter_selection_treatment_providers_spec}
\title{Filter selection for trearment providers specification}
\format{
An object of class \code{character} of length 15.
}
\usage{
filter_selection_treatment_providers_spec
}
\description{
Filter selection for trearment providers specification
}
\keyword{datasets}
