% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataFilterSelection.R
\docType{data}
\name{filter_selection_zip}
\alias{filter_selection_zip}
\title{Filter selection for zip}
\format{
An object of class \code{character} of length 47.
}
\usage{
filter_selection_zip
}
\description{
Filter selection for zip
}
\keyword{datasets}
