% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{FOCAL_DATA_DICTIONARY}
\alias{FOCAL_DATA_DICTIONARY}
\title{FOCAL DATA DICTIONARY}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 10 columns.
}
\usage{
FOCAL_DATA_DICTIONARY
}
\description{
FOCAL DATA DICTIONARY
}
\keyword{datasets}
