% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{FOCAL_USER_TABLE}
\alias{FOCAL_USER_TABLE}
\title{FOCAL USER TABLE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 6 columns.
}
\usage{
FOCAL_USER_TABLE
}
\description{
FOCAL USER TABLE
}
\keyword{datasets}
