% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-MapSf.R
\docType{data}
\name{FRANKLIN_COUNTY_SCHOOL_DISTRICT_SF_PATH}
\alias{FRANKLIN_COUNTY_SCHOOL_DISTRICT_SF_PATH}
\title{The path of shapefiles of school districts in Franklin County}
\format{
An object of class \code{character} of length 1.
}
\usage{
FRANKLIN_COUNTY_SCHOOL_DISTRICT_SF_PATH
}
\description{
The path of shapefiles of school districts in Franklin County
}
\keyword{datasets}
