% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map_sf.R
\docType{data}
\name{franklin_county_school_districts}
\alias{franklin_county_school_districts}
\title{Array of the names of Franklin county school districts}
\format{
An object of class \code{character} of length 17.
}
\usage{
franklin_county_school_districts
}
\description{
Array of the names of Franklin county school districts
}
\keyword{datasets}
