% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geocoding_sf}
\alias{geocoding_sf}
\title{Geolocate a point among the provided sf objects}
\usage{
geocoding_sf(lat, lng, sf, id = "name")
}
\arguments{
\item{lat}{lat}

\item{lng}{lng}

\item{sf}{a sf tibble that contains many sf objects}

\item{id}{the identifier column of the sf object}
}
\value{
the name of the object that the point falls into; return NA if none
}
\description{
Geolocate a point among the provided sf objects
}
