% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_census_tract_sf}
\alias{get_census_tract_sf}
\title{Load census tract shapefiles}
\usage{
get_census_tract_sf(parquet = TRUE)
}
\arguments{
\item{parquet}{parquet format}
}
\value{
sf object
}
\description{
Load census tract shapefiles
}
\examples{
get_census_tract_sf()
}
