% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_cota_bus_lines_sf}
\alias{get_cota_bus_lines_sf}
\title{Load COTA bus lines shapefiles}
\usage{
get_cota_bus_lines_sf(parquet = TRUE)
}
\arguments{
\item{parquet}{parquet format}
}
\value{
sf object
}
\description{
Load COTA bus lines shapefiles
}
