% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapefile_loading_function.R
\name{get_fire_district_sf}
\alias{get_fire_district_sf}
\title{Load fire district shapefiles}
\usage{
get_fire_district_sf()
}
\value{
sf object
}
\description{
Load fire district shapefiles
}
\examples{
get_fire_district_sf()
}
