% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-hotSpotDetection.R
\name{get_hot_spot_business}
\alias{get_hot_spot_business}
\title{Get the eligible business entities in a hot spot region}
\usage{
get_hot_spot_business(
  hot_spot_id,
  lat_min,
  lat_max,
  lng_min,
  lng_max,
  business_location,
  eligible_business_type
)
}
\arguments{
\item{hot_spot_id}{the id of the hot spot}

\item{lat_min}{min lat}

\item{lat_max}{max lat}

\item{lng_min}{min lng}

\item{lng_max}{max lng}

\item{business_location}{table of business location}

\item{eligible_business_type}{eligible business type and category}
}
\value{
table of eligible business
}
\description{
Get the eligible business entities in a hot spot region
}
