% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_spot_detection.R
\name{get_hot_spot_region}
\alias{get_hot_spot_region}
\title{Get the hot spot region from spatial data}
\usage{
get_hot_spot_region(
  od_data = opioidDashboard::opioid_overdose_data(),
  percent_tile = 0.75,
  bin_width = c(0.01, 0.01)
)
}
\arguments{
\item{od_data}{opioid overdose data}

\item{percent_tile}{hot spot threshold}

\item{bin_width}{size of the bin, a 2-d (x, y) vector}
}
\value{
a tibble with hot spot spatial info
}
\description{
Get the hot spot region from spatial data
}
\examples{
# Change the od_data to get different hot spot region
}
