% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_naloxone_box_sf}
\alias{get_naloxone_box_sf}
\title{Load naloxone box shapefiles}
\usage{
get_naloxone_box_sf(parquet = TRUE)
}
\arguments{
\item{parquet}{parquet format}
}
\value{
sf object
}
\description{
Load naloxone box shapefiles
}
