% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-ohioBusiness.R
\name{get_ohio_business}
\alias{get_ohio_business}
\title{Get the business information from Ohio}
\usage{
get_ohio_business()
}
\value{
business information from Ohio
}
\description{
Get the business information from Ohio
}
