% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_ohio_school_district_sf}
\alias{get_ohio_school_district_sf}
\title{Load Ohio school district shapefiles}
\usage{
get_ohio_school_district_sf(parquet = TRUE)
}
\arguments{
\item{parquet}{parquet format}
}
\value{
sf object
}
\description{
Load Ohio school district shapefiles
}
\examples{
get_ohio_school_district_sf()
}
