% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-opioidOverdoseRate.R
\name{get_region_od_rate}
\alias{get_region_od_rate}
\title{Get the overdose rate of the regions in a sf object}
\usage{
get_region_od_rate(
  region = c("zip", "school_district", "EMS", "census_tract", "fire_district"),
  region_sf,
  id_colname,
  label_colname,
  pop_colname,
  od_data = opioidDashboard::opioid_overdose_data()
)
}
\arguments{
\item{region}{"zip", "school_district", "EMS", "census_tract"}

\item{region_sf}{sf object}

\item{id_colname}{unique identifier column name}

\item{label_colname}{region label column name}

\item{pop_colname}{population column name}

\item{od_data}{overdose data}
}
\value{
sf object with \code{n} and \code{rate}
}
\description{
Get the overdose rate of the regions in a sf object
}
