% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapefile_loading_function.R
\name{get_zipcode_sf}
\alias{get_zipcode_sf}
\title{Load zip code level shapefiles}
\usage{
get_zipcode_sf()
}
\value{
sf object
}
\description{
Load zip code level shapefiles
}
\examples{
get_zipcode_sf()
}
