% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naloxone_data_loading_functions.R
\name{kit_data}
\alias{kit_data}
\title{Load kit data extracted from naloxone intake form}
\usage{
kit_data(dir = opioidDashboard::NALOXONE_DATA_DIRECTORY)
}
\arguments{
\item{dir}{data directory on OSC that contains all of the naloxone intake forms data}
}
\value{
kit data
}
\description{
Load kit data extracted from naloxone intake form
}
\examples{
kit_data()
}
