% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_directory_doc.R
\docType{data}
\name{NALOXONE_DATA_DIRECTORY}
\alias{NALOXONE_DATA_DIRECTORY}
\title{The data directory that contains all of the cleaned
naloxone intake forms data cleaned fromt the opioidAction
data preparation package}
\format{
An object of class \code{character} of length 1.
}
\usage{
NALOXONE_DATA_DIRECTORY
}
\description{
The data directory that contains all of the cleaned
naloxone intake forms data cleaned fromt the opioidAction
data preparation package
}
\keyword{datasets}
