% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naloxone_forms_data_loading_functions.R
\name{naloxone_usage_data}
\alias{naloxone_usage_data}
\title{Load naloxone usage data extracted from naloxone intake form}
\usage{
naloxone_usage_data(dir = opioidDashboard::NALOXONE_DATA_DIRECTORY)
}
\arguments{
\item{dir}{data directory on OSC that contains all of the naloxone intake forms data}
}
\value{
naloxone usage data
}
\description{
Load naloxone usage data extracted from naloxone intake form
}
\examples{
naloxone_usage_data()
}
