% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-overdoseDataPalette.R
\docType{data}
\name{od_case_rate_init_pal_list}
\alias{od_case_rate_init_pal_list}
\title{overdose case rate initial palette function}
\format{
An object of class \code{list} of length 3.
}
\usage{
od_case_rate_init_pal_list
}
\description{
overdose case rate initial palette function
}
\keyword{datasets}
