% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{OHIO_BUSINESS_DATA_DIRECTORY}
\alias{OHIO_BUSINESS_DATA_DIRECTORY}
\title{Ohio business information}
\format{
An object of class \code{character} of length 1.
}
\usage{
OHIO_BUSINESS_DATA_DIRECTORY
}
\description{
Ohio business information
}
\keyword{datasets}
