% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map_sf.R
\docType{data}
\name{OHIO_SCHOOL_DISTRICT_SF_PATH}
\alias{OHIO_SCHOOL_DISTRICT_SF_PATH}
\title{The path of shapefiles of all of the Ohio's school districts}
\format{
An object of class \code{character} of length 1.
}
\usage{
OHIO_SCHOOL_DISTRICT_SF_PATH
}
\description{
The path of shapefiles of all of the Ohio's school districts
}
\keyword{datasets}
