% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_directory_doc.R
\docType{data}
\name{OPIOID_OVERDOSE_DATA_DIRECTORY}
\alias{OPIOID_OVERDOSE_DATA_DIRECTORY}
\title{Contains the data directory of opioid overdose data}
\format{
An object of class \code{character} of length 1.
}
\usage{
OPIOID_OVERDOSE_DATA_DIRECTORY
}
\description{
Contains the data directory of opioid overdose data
}
\keyword{datasets}
