% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map_init_bounds.R
\docType{data}
\name{opioid_overdose_map_init_bounds}
\alias{opioid_overdose_map_init_bounds}
\title{initial bounds for opioid ovedose map
a list object
list(
min_lng = min(od_data$lng, na.rm = TRUE),
min_lat = min(od_data$lat, na.rm = TRUE),
max_lng = max(od_data$lng, na.rm = TRUE),
max_lat = max(od_data$lat, na.rm = TRUE)
)}
\format{
An object of class \code{list} of length 4.
}
\usage{
opioid_overdose_map_init_bounds
}
\description{
initial bounds for opioid ovedose map
a list object
list(
min_lng = min(od_data$lng, na.rm = TRUE),
min_lat = min(od_data$lat, na.rm = TRUE),
max_lng = max(od_data$lng, na.rm = TRUE),
max_lat = max(od_data$lat, na.rm = TRUE)
)
}
\keyword{datasets}
