% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-overdoseDataPalette.R
\docType{data}
\name{overdose_case_rate_palette}
\alias{overdose_case_rate_palette}
\title{overdose case rate default palette}
\format{
An object of class \code{character} of length 1.
}
\usage{
overdose_case_rate_palette
}
\description{
overdose case rate default palette
}
\keyword{datasets}
