% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naloxone_forms_data_loading_functions.R
\name{personal_data}
\alias{personal_data}
\title{Load personal data extracted from naloxone intake form}
\usage{
personal_data(dir = opioidDashboard::NALOXONE_DATA_DIRECTORY)
}
\arguments{
\item{dir}{data directory on OSC that contains all of the naloxone intake forms data}
}
\value{
personal data
}
\description{
Load personal data extracted from naloxone intake form
}
\examples{
personal_data()
}
