% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-naloxoneData.R
\name{program_data}
\alias{program_data}
\title{Load program data extracted from naloxone intake form}
\usage{
program_data(dir = opioidDashboard::NALOXONE_DATA_DIRECTORY)
}
\arguments{
\item{dir}{data directory on OSC that contains all of the naloxone intake forms data}
}
\value{
program data
}
\description{
Load program data extracted from naloxone intake form
}
\examples{
program_data()
}
