% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapefile_loading_function.R,
%   R/naloxone_data_loading_functions.R
\name{project_dawn_focused_indicators}
\alias{project_dawn_focused_indicators}
\title{Load focused indicators  extracted from naloxone intake form}
\usage{
project_dawn_focused_indicators(dir = opioidDashboard::NALOXONE_DATA_DIRECTORY)

project_dawn_focused_indicators(dir = opioidDashboard::NALOXONE_DATA_DIRECTORY)
}
\arguments{
\item{dir}{data directory on OSC that contains all of the naloxone intake forms data}
}
\value{
focused naloxone intake forms indicators

focused naloxone intake forms indicators
}
\description{
Load focused indicators  extracted from naloxone intake form

Load focused indicators  extracted from naloxone intake form
}
\examples{
project_dawn_focused_indicators()
project_dawn_focused_indicators()
}
