% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-other.R
\docType{data}
\name{PUBLIC_PLACES_ICON_LIST}
\alias{PUBLIC_PLACES_ICON_LIST}
\title{PUBLIC_PLACES_ICON_LIST}
\format{
An object of class \code{list} of length 10.
}
\usage{
PUBLIC_PLACES_ICON_LIST
}
\description{
PUBLIC_PLACES_ICON_LIST
}
\keyword{datasets}
