% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{ROOT_PATH}
\alias{ROOT_PATH}
\title{The root directory of FOCAL data ingestion pipeline}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROOT_PATH
}
\description{
The root directory of FOCAL data ingestion pipeline
}
\keyword{datasets}
