% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{schema_naloxone}
\alias{schema_naloxone}
\title{naloxone schema used for data loading}
\format{
An object of class \code{list} of length 23.
}
\usage{
schema_naloxone
}
\description{
naloxone schema used for data loading
}
\keyword{datasets}
