% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{schema_primary}
\alias{schema_primary}
\title{primary table schema used for data loading}
\format{
An object of class \code{list} of length 26.
}
\usage{
schema_primary
}
\description{
primary table schema used for data loading
}
\keyword{datasets}
