% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-treatmentProviders.R
\name{treatment_providers_data}
\alias{treatment_providers_data}
\title{Load treament providers data from OSC}
\usage{
treatment_providers_data(
  dir = opioidDashboard::TREATMENT_PROVIDERS_DATA_DIRECTORY,
  tier_2_only = TRUE,
  processed = TRUE,
  parquet = FALSE
)
}
\arguments{
\item{dir}{directory on OSC that contains treament providers data}

\item{tier_2_only}{only show tier 2 treatment providers}

\item{processed}{processed data}

\item{parquet}{parquet}
}
\value{
treatment providers data
}
\description{
Load treament providers data from OSC
}
