% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc-dataDirectory.R
\docType{data}
\name{TREATMENT_PROVIDERS_DATA_DIRECTORY}
\alias{TREATMENT_PROVIDERS_DATA_DIRECTORY}
\title{Treatment providers data directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
TREATMENT_PROVIDERS_DATA_DIRECTORY
}
\description{
Treatment providers data directory
}
\keyword{datasets}
