% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_directory_doc.R
\docType{data}
\name{DRUG_CRIME_DATA_DIRECTORY}
\alias{DRUG_CRIME_DATA_DIRECTORY}
\title{Drug crime data directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
DRUG_CRIME_DATA_DIRECTORY
}
\description{
Drug crime data directory
}
\keyword{datasets}
