% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overdose_filter_selection.R
\docType{data}
\name{filter_selection_agency}
\alias{filter_selection_agency}
\title{Filter selection for agency}
\format{
An object of class \code{character} of length 13.
}
\usage{
filter_selection_agency
}
\description{
Filter selection for agency
}
\keyword{datasets}
