% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapefile_loading_function.R
\name{get_census_tract_sf}
\alias{get_census_tract_sf}
\title{Load census tract shapefiles}
\usage{
get_census_tract_sf()
}
\value{
sf object
}
\description{
Load census tract shapefiles
}
\examples{
get_census_tract_sf()
}
