% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_city_sf}
\alias{get_city_sf}
\title{Load City, Village, and Township outlines}
\usage{
get_city_sf(root = opioidDashboard::ROOT_PATH)
}
\value{
sf object
}
\description{
Load City, Village, and Township outlines
}
