% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_fire_district_sf}
\alias{get_fire_district_sf}
\title{Load fire district shapefiles}
\usage{
get_fire_district_sf(parquet = TRUE)
}
\arguments{
\item{parquet}{parquet format}
}
\value{
sf object
}
\description{
Load fire district shapefiles
}
\examples{
get_fire_district_sf()
}
