% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapefile_loading_function.R
\name{get_franklin_county_school_district_sf}
\alias{get_franklin_county_school_district_sf}
\title{Load Franklin county school district shapefiles}
\usage{
get_franklin_county_school_district_sf()
}
\value{
sf object
}
\description{
Load Franklin county school district shapefiles
}
\examples{
get_franklin_county_school_district_sf()
}
