% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_jurisdictions_sf}
\alias{get_jurisdictions_sf}
\title{Load Jurisdictions}
\usage{
get_jurisdictions_sf(parquet = TRUE)
}
\arguments{
\item{parquet}{parquet format}
}
\value{
sf object
}
\description{
Load Jurisdictions
}
