% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapefile_loading_function.R
\name{get_naloxone_box_sf}
\alias{get_naloxone_box_sf}
\title{Load naloxone box shapefiles}
\usage{
get_naloxone_box_sf()
}
\value{
sf object
}
\description{
Load naloxone box shapefiles
}
