% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-mapShapefile.R
\name{get_response_area_sf}
\alias{get_response_area_sf}
\title{Load response area shapefiles}
\usage{
get_response_area_sf()
}
\value{
sf object
}
\description{
Load response area shapefiles
}
\examples{
get_response_area_sf()
}
